# Reduce time of execution when requesting totalCount with stockAvailability filter